feat: first iteration of connection + connectionPoint logic
Basic idea: Connections in the planner happen between connection points. For
the current connections, these happen to be part of resources, but they can
live anywhere.  The connection is then established as the path between
ConnectionPoint#from and ConnectionPoint#to, as determined by their IDs
(strings).

It is currently a little limited / rough, since the DOM actually does not change
for the points themselves. It is therefore a little tricky to only invalidate the
necessary part of the tree.